+++ test/process_cluster_diff_ports_test.go
@@ -30,8 +30,7 @@ import (
 
 func TestProcessClusterDifferentPorts(t *testing.T) {
 	removeArangodProcesses(t)
-	needTestMode(t, testModeProcess)
-	needStarterMode(t, starterModeCluster)
+	testMatch(t, testModeProcess, starterModeCluster, false)
 	dataDirMaster := SetUniqueDataDir(t)
 	defer os.RemoveAll(dataDirMaster)